Improve city detection speed with R-Tree
This allows to quickly find the city in which
the point may be located. Unfortunately,
while this R-Tree greatly speeds up finding the city
by about 2000x, testing whether the point is
actually in the city is now the bottleneck.

Either some additional caching layer or
a different point-in-polygon testing algorithm
patched into the geo crate will be necessary
to further improve performance.